requirements.txt
asgiref==3.4.1
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
credentials==1.1
Django==4.0
django-countries==7.2.1
django-debug-toolbar==3.2.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
firebase==3.0.1
firebase-admin==5.2.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
msgpack==1.0.3
packaging==21.3
Pillow==8.4.0
proto-plus==1.19.8
protobuf==3.19.1
psycopg==3.0.7
psycopg2==2.9.2
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
